requirements.txt
pandas==2.2.3
numpy==2.2.0
pyarrow==19.0.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
plotly==5.22.0
kaleido==0.2.1
python-dotenv==1.0.1
faker==25.0.1